+++ libs/react-ui/src/index.ts
@@ -4,4 +4,4 @@ export {
   ButtonSize,
 } from './lib/button/button';
 
-export { default as Tooltip } from './lib/Tooltip';
+export { default as Tooltip } from './lib/Tooltip-rename';
